delivery can add new address

+++ Documents/NetBeansProjects/JLS/src/jls/Kun/Fresh.java
@@ -29,11 +29,11 @@ public class Fresh extends javax.swing.JPanel {
     public class Item{
     public String item;
     public boolean order;
-    public Double price;
+    public int price;
     public int quantity;
-    public Double total;
+    public int total;
         
-        public Item(String Item, boolean Order, Double Price, int Quantity, Double Total){
+        public Item(String Item, boolean Order, int Price, int Quantity, int Total){
             this.item = Item;
             this.order = Order;
             this.price = Price;
@@ -45,8 +45,8 @@ public class Fresh extends javax.swing.JPanel {
     public ArrayList ListItem(){
         ArrayList<Item> list = new ArrayList<Item>();
         
-        Item flower = new Item("Rose", false, 20.00, 0, 0.00);
-        Item flower1 = new Item("Lily", false, 30.00, 0, 0.00);
+        Item flower = new Item("Rose", false, 20, 0, 0);
+        Item flower1 = new Item("Lily", false, 30, 0, 0);
         list.add(flower);
         list.add(flower1);
         return list;
@@ -88,7 +88,7 @@ public class Fresh extends javax.swing.JPanel {
             }
         ) {
             Class[] types = new Class [] {
-                java.lang.Object.class, java.lang.Boolean.class, java.lang.Double.class, java.lang.Integer.class, java.lang.Double.class
+                java.lang.Object.class, java.lang.Boolean.class, java.lang.Integer.class, java.lang.Integer.class, java.lang.Integer.class
             };
             boolean[] canEdit = new boolean [] {
                 false, true, false, true, false
